use CanvasLayer
