+++ src/lib/marks/helpers/CanvasLayer.svelte
@@ -1,17 +1,34 @@
 <script lang="ts">
+    import { getContext } from 'svelte';
+    import type { PlotContext } from 'svelteplot/types';
+
     let {
-        canvas = $bindable(),
         devicePixelRatio = $bindable(1),
-        plot
+        ...restProps
     }: {
-        canvas: HTMLCanvasElement;
         devicePixelRatio: number;
-        plot: PlotState;
     } = $props();
 
+    const { getPlotState } = getContext<PlotContext>('svelteplot');
+    const plot = $derived(getPlotState());
+
+    // code from https://developer.mozilla.org/en-US/docs/Web/API/Window/devicePixelRatio
+    let remove: null | (() => void) = null;
+
+    function updatePixelRatio() {
+        if (remove != null) {
+            remove();
+        }
+        const mqString = `(resolution: ${window.devicePixelRatio}dppx)`;
+        const media = matchMedia(mqString);
+        media.addEventListener('change', updatePixelRatio);
+        remove = () => {
+            media.removeEventListener('change', updatePixelRatio);
+        };
+        devicePixelRatio = window.devicePixelRatio;
+    }
     $effect(() => {
-        devicePixelRatio = window.devicePixelRatio || 1;
-        const ctx = canvas.getContext('2d');
+        updatePixelRatio();
     });
 </script>
 
@@ -24,7 +41,7 @@
 <foreignObject x="0" y="0" width={plot.width} height={plot.height}>
     <canvas
         xmlns="http://www.w3.org/1999/xhtml"
-        bind:this={canvas}
+        {...restProps}
         width={plot.width * devicePixelRatio}
         height={plot.height * devicePixelRatio}
         style="width: {plot.width}px; height: {plot.height}px;"></canvas>
